First commit
Does not currently work, but has most of the infrastructure for:
- Creating new networks (Empty and random)
- "Loading" a network based on a set of weights
- Input, output and hidden layers, and the nodes thereof

Commit log for later:
Added TODO

I used the weight initialisation function Xavier Weight Initialization, as described in the paper "Understanding the difficulty of training deep feedforward neural networks" by Xavier Glorot. Although this is an old paper, I have seen this initialisation function used in the present day. It also functions well with the sigmoid activation function which I plan on using.